requirements.txt
# API Clients
requests==2.31.0
tmdbsimple==2.9.1

# Data Processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.0

# Deep Learning (Phase 2)
torch==2.0.1
tensorflow==2.13.0

# Web Framework
flask==2.3.3
flask-cors==4.0.0

# Database
sqlalchemy==2.0.20
psycopg2-binary==2.9.7

# Development
python-dotenv==1.0.0
pytest==7.4.0
black==23.7.0
isort==5.12.0

# Visualization
matplotlib==3.7.2
seaborn==0.12.2 